requirements.txt
altgraph==0.17.4
customtkinter==5.2.2
darkdetect==0.8.0
packaging==25.0
pefile==2023.2.7
pillow==11.2.1
pyinstaller==6.14.0
pyinstaller-hooks-contrib==2025.4
pywin32-ctypes==0.2.3
setuptools==80.9.0
